[cm] Create wavelets file
